Clicking on own name will direct to pathname /user/userID
When a user clicks on one's own name to view profile, the url path will now include the user id in the pathname. For example, if user #21 would click on his own name to view his profile, then the pathname would be .../user/21. However, the contents of the profile page are still static, thus if changing the user ID in the pathname, then the information displayed will still be that of the logged in user.