User ID displayed in profile Page
When navigating to user profiles, the profile page displays the ID according to the URL. Thus, if accessing URL ../user/12, then the profile page will display "user id: 12". However, the other parameters such as name, last name, email and engineering discipline are still static and do not change as they are props of the logged in user.