User Profile Page Routing
The profile information changes dynamically according to the user ID in the url. When redirecting to the path of a specific user, the page will fetch the user's profile information based on the id in the url. Still requires formatting, as well as a "Page not Found" for inexistent user IDs.