Additional bug fixes to profile and ID added
When viewing another user's profile and then clicking on one's own name to view one's own profile, this would redirec to the correct url but it would not update the information. As a result, the link that redirects to one's own profile now includes an onclick event that reloads the page. This will allow for the re-rendering of the profile. Moreover, the user ID is now displayed on the user profile.