Answer is seen after submitting
When viewing a question, a logged in user can provide an answer and once the answer is submitted, the component will be reloaded to display the answer provided. In other words, once an answer is submitted, the page will not have to be reloaded to see the answer. However, at the moment, a user can only submit a single answer before having to refresh the page.