Editing user information
User can now change first name, last name, email, engineering and password. There is no verification before changing name, email and engineering. However, for changing password, the email and current password must be entered and verified. Alas, checking for the email is redundant is that information is available on the user profile. Additionally, the new password will also be validated.