Edit Profile Form
Instead of separate modals, editing information is done using a single a form. This form is found at the path /edit. The submit button is yet to be functional, but the cancel button returns a user to their profile. The profile page has yet to display a button which leads to the form though, so at the moment one would have to type the path into the URL.